chore: Split `Nargo.toml` operations into separate package
