+++ crates/nargo_cli/Cargo.toml
@@ -24,6 +24,7 @@ dirs.workspace = true
 url.workspace = true
 iter-extended.workspace = true
 nargo.workspace = true
+nargo_toml.workspace = true
 noir_lsp.workspace = true
 noirc_driver.workspace = true
 noirc_frontend.workspace = true
