+++ crates/nargo_cli/src/cli/compile_cmd.rs
@@ -4,6 +4,7 @@ use iter_extended::try_vecmap;
 use iter_extended::vecmap;
 use nargo::package::Package;
 use nargo::{artifacts::contract::PreprocessedContract, NargoError};
+use nargo_toml::{find_package_manifest, resolve_workspace_from_toml};
 use noirc_driver::{
     compile_contracts, compile_main, CompileOptions, CompiledProgram, ErrorsAndWarnings, Warnings,
 };
@@ -15,8 +16,7 @@ use clap::Args;
 use nargo::ops::{preprocess_contract_function, preprocess_program};
 
 use crate::errors::{CliError, CompileError};
-use crate::manifest::resolve_workspace_from_toml;
-use crate::{find_package_manifest, prepare_package};
+use crate::prepare_package;
 
 use super::fs::{
     common_reference_string::{
