+++ crates/nargo_cli/src/cli/mod.rs
@@ -1,10 +1,9 @@
 use clap::{Args, Parser, Subcommand};
 use const_format::formatcp;
+use nargo_toml::find_package_root;
 use std::path::PathBuf;
 
 use color_eyre::eyre;
-
-use crate::find_package_root;
 
 mod fs;
