+++ crates/nargo_toml/Cargo.toml
@@ -1,0 +1,20 @@
+[package]
+name = "nargo_toml"
+description = "Utilities for working with Nargo.toml files"
+version.workspace = true
+authors.workspace = true
+edition.workspace = true
+
+# See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
+
+[dependencies]
+dirs.workspace = true
+fm.workspace = true
+nargo.workspace = true
+noirc_frontend.workspace = true
+serde.workspace = true
+thiserror.workspace = true
+toml.workspace = true
+url.workspace = true
+
+[dev-dependencies]
