+++ crates/nargo_toml/src/lib.rs
@@ -1,5 +1,6 @@
 use std::{
     collections::BTreeMap,
+    fs::ReadDir,
     path::{Path, PathBuf},
 };
 
@@ -11,7 +12,55 @@ use nargo::{
 use noirc_frontend::graph::CrateName;
 use serde::Deserialize;
 
-use crate::{errors::ManifestError, git::clone_git_repo};
+mod errors;
+mod git;
+
+pub use errors::ManifestError;
+use git::clone_git_repo;
+
+/// Returns the path of the root directory of the package containing `current_path`.
+///
+/// Returns a `CliError` if no parent directories of `current_path` contain a manifest file.
+pub fn find_package_root(current_path: &Path) -> Result<PathBuf, ManifestError> {
+    let manifest_path = find_package_manifest(current_path)?;
+
+    let package_root =
+        manifest_path.parent().expect("infallible: manifest file path can't be root directory");
+
+    Ok(package_root.to_path_buf())
+}
+
+/// Returns the path of the manifest file (`Nargo.toml`) of the package containing `current_path`.
+///
+/// Returns a `CliError` if no parent directories of `current_path` contain a manifest file.
+pub fn find_package_manifest(current_path: &Path) -> Result<PathBuf, ManifestError> {
+    current_path
+        .ancestors()
+        .find_map(|dir| find_file(dir, "Nargo", "toml"))
+        .ok_or_else(|| ManifestError::MissingFile(current_path.to_path_buf()))
+}
+
+// Looks for file named `file_name` in path
+fn find_file<P: AsRef<Path>>(path: P, file_name: &str, extension: &str) -> Option<PathBuf> {
+    let entries = list_files_and_folders_in(path)?;
+    let file_name = format!("{file_name}.{extension}");
+
+    find_artifact(entries, &file_name)
+}
+
+// There is no distinction between files and folders
+fn find_artifact(entries: ReadDir, artifact_name: &str) -> Option<PathBuf> {
+    let entry = entries
+        .into_iter()
+        .flatten()
+        .find(|entry| entry.file_name().to_str() == Some(artifact_name))?;
+
+    Some(entry.path())
+}
+
+fn list_files_and_folders_in<P: AsRef<Path>>(path: P) -> Option<ReadDir> {
+    std::fs::read_dir(path).ok()
+}
 
 #[derive(Debug, Deserialize, Clone)]
 struct PackageConfig {
@@ -277,7 +326,7 @@ fn resolve_package_from_toml(toml_path: &Path) -> Result<Package, ManifestError>
 }
 
 /// Resolves a Nargo.toml file into a `Workspace` struct as defined by our `nargo` core.
-pub(crate) fn resolve_workspace_from_toml(
+pub fn resolve_workspace_from_toml(
     toml_path: &Path,
     selected_package: Option<CrateName>,
 ) -> Result<Workspace, ManifestError> {
